@@ -8,20 +8,20 @@ repository = "https://github.com/poofeg/vk-to-commerceml"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.11"
-pydantic = "^2.5.3"
-pydantic-xml = {version = "^2.8.1", extras = ["lxml"]}
-pydantic-settings = "^2.2.1"
-typer = {extras = ["all"], version = "^0.9.0"}
-aiohttp = {extras = ["aiodns"], version = "^3.9.3"}
-fastapi = "^0.110.0"
-uvicorn = {extras = ["standard"], version = "^0.27.1"}
-aiogram = {extras = ["redis"], version = "^3.4.1"}
-cryptography = "^42.0.5"
+python = "^3.13"
+pydantic = "^2.9.2"
+pydantic-xml = {version = "^2.14.0", extras = ["lxml"]}
+pydantic-settings = "^2.6.1"
+typer = {version = "^0.13.1", extras = ["all"]}
+aiohttp = {version = "^3.10.11", extras = ["speedups"]}
+fastapi = "^0.115.5"
+uvicorn = {version = "^0.32.1", extras = ["standard"]}
+aiogram = {version = "^3.15.0", extras = ["redis"]}
+cryptography = "^43.0.3"
 
 [tool.poetry.group.dev.dependencies]
-mypy = {version = "^1.9.0", extras = ["dmypy"]}
-types-aiofiles = "^23.2.0.20240311"
+mypy = {version = "^1.13.0", extras = ["dmypy"]}
+types-aiofiles = "^24.1.0.20240626"
 
 [build-system]
 requires = ["poetry-core"]
